feat: add support for non-ASCII tags, authors, and categories
For example, now Chinese can be used in the names of tags, authors, and categories.

Previously, this was blocked due to bugs in Hugo.

As of Hugo v0.65, new functionality for taxonomies was introduced. The Hugo taxonomy feature powers authors, categories, and tags in Academic.

BREAKING CHANGE:
Requires Hugo v0.65+